gui title, corner prefabs, fixing mites, market

+++ src/main/kotlin/dev/dungeoncrawler/data/MarketData.kt
@@ -3,4 +3,4 @@ package dev.dungeoncrawler.data
 import org.bukkit.inventory.ItemStack
 import java.util.*
 
-class MarketData(var items: ArrayList<Triple<Date, Pair<Double, Boolean>, ItemStack>>? = null)
+class MarketData(var items: ArrayList<Triple<Date, Pair<Int, Boolean>, ItemStack>>? = null)
